data_analysis/requirements.txt
fastapi
uvicorn
pymongo
pandas
python-dotenv
scikit-learn